Replace Vectors with SIMDs
VectorNs are now type aliases of SIMDN<Float> with an extension to add
more methods such as dot/cross products, interpolation, etc.
This allows to use the pre-existing methods of the SIMD protocol and
should improve performance.

The major drawback is that now Matrices are back to being arrays of
VectorNs.